clip improvements
use -maxrate instead of -b to ensure it doesnt go over and it can go
under

only scale when over 1080p. now that we got 8 megs we might as well use
theonly scale when over 1080p. now that we got 8 megs we might as well
use them

force pix_fmt for better compat